[Pdm2-11] Move Campaings inside DM directory
