+++ src/app/(Restricted)/(DM)/Campaigns/components/AddCampaign.tsx
@@ -1,6 +1,6 @@
 import { View } from "react-native";
 import React from "react";
-import Button from "../../../../common/components/Button";
+import Button from "../../../../../common/components/Button";
 
 export default function AddCampaign() {
   return (
