+++ src/app/_navigation/DM/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CampaignsScreen from "../../(Restricted)/Campaigns/screens/CampaignsScreen";
+import CampaignsScreen from "../../(Restricted)/(DM)/Campaigns/screens/CampaignsScreen";
 
 export default function DmIndex() {
   return <CampaignsScreen />;
